Rename docker types to container
Docker has asked for packages beginning with the name docker be removed from
github.  We want to follow their requests and since we plan on using this
policy for other container runtimes, we need to rename the types to something
more generic.

We will need to be careful when updating docker.pp based systems to container.pp.