container_runtime_t Rules needed to work with unconfined module disabled.
When you disable the unconfined module, lots of rules are removed from container
runtime policy which also removes it from container_t, since it is typebounds.

Adding these rules to container_runtime makes it more powerfull, but we ususally
run it unconfined so it should not be a problem. Bottom line container_runtime_t
has to be a superset of allow rules for container_t types.