Allow containers to show IPC information using `ipcs`
Containers using the `ipcs` command included in busybox fail unless this
rule is included.

The implementation of `ipcs` in busybox uses shmctl(0, SHM_INFO, ...)
which needs to be allowed for it to work.
https://git.busybox.net/busybox/tree/util-linux/ipcs.c?h=1_28_0#n138

(The implementation of `ipcs` in util-linux doesn't suffer from this issue,
it defaults to trying to read from the /proc files and only uses shmctl
commands as a fallback:
https://github.com/karelzak/util-linux/blob/v2.27.1/sys-utils/ipcutils.c#L108)

Tested:

- Before this change:

    $ docker run busybox ipcs
    kernel not configured for message queues
    kernel not configured for shared memory
    kernel not configured for semaphores

- After this change:

    $ docker run busybox ipcs
    ------ Message Queues --------
    key        msqid      owner      perms      used-bytes   messages
    ------ Shared Memory Segments --------
    key        shmid      owner      perms      bytes      nattch     status
    ------ Semaphore Arrays --------
    key        semid      owner      perms      nsems

Signed-off-by: Filipe Brandenburger <filbranden@google.com>