remove k3s fcontexts
K3s brings its own embedded userland runtime and dumps it on disk under
/var/lib/rancher/k3s/data/${content_id=-"current"}/ which includes runc,
containerd, and the containerd shims. The initial k3s policy implementation
re-used existing policy types from container-selinux and marked all of
the disgorged userland (binaries and config, alike) as
container_runtime_exec_t which is what runc and the containerd bits
require. Because we ship busybox in this userland, with the shell
applet, this can lead to elevated privileges for any unconfined seuser
(the default for unmodified installations).

To address this problem we have re-written the k3s-selinux policy to
declare and leverage a couple of custom types for the various bits of
"content" that k3s brings. As such, the k3s-specific file-contexts that
were pushed upstream to here with containers/container-selinux#140 and
recently modified with containers/container-selinux#156 are no longer
appropriate to be maintained here. This PR removes the k3s-specific
file-context entries for:
- systemd units
- k3s binaries under /usr and /usr/local
- everything under /var/lib/rancher/k3s and /var/run/k3s

Signed-off-by: Jacob Blain Christen <jacob@rancher.com>