Allow access to BPF Filesystems
BPF Filesystems (bpffs) are used to manage BPF Object Lifecycles.
For example, you can place a special object (called a pin) on a BPF
filesystem and use that object to get a reference to the
program/link/map that it references if you have permissions to make
bpf() syscalls.

In a recent kernel patch, a container may also recieve a token
that allows it to make BPF syscalls on this filesystem.

Since bpffs doesn't support xattrs, it's not possible to bind mount
it from the host with `:Z` so it can be re-labelled. Adding xattrs
to bpffs is not something that I've looked into in detail, but I
imagine it's non-trivial.

This commit should add the necessary rules to ensure that SELinux
permits the use of bpf_t file/directory objects from containers
when the BPFFS volume was bind mounted from the host.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>