Rename all /var/run file context entries to /run
With the 1f76e522a ("Rename all /var/run file context entries to /run")
selinux-policy commit, all /var/run file context entries moved to /run
and the equivalency was inverted. Subsequently, changes in container.fc
need to be done, too, in a similar manner. Together with that, the
"/run/lock = /var/lock" equivalency was also inverted and entries moved.

For RHEL < 10 and Fedora < 40, these changes are reverted in specfile.

Signed-off-by: Zdenek Pytela <zpytela@redhat.com>