change label of kubelet\'s socket to kubernetes_file_t
This commit is a continuation of https://github.com/containers/container-selinux/pull/275/files
When Kubelet starts it creates /var/lib/kubelet/pod-resources/kubelet.sock file on the fly.
This means the socket file inheret its type from the parent so it still gets the container_var_lib_t type.

We're adding the container_lib_filetrans in order to change the socket to the correct label.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>