TMT: simplify podman tests
`dnf download` on CentOS Stream ends up downloading all rpm versions of
a package from all available repos instead of only the latest rpm. This
leads to more (unnecessary) complications in the test script.

Things are a lot simpler if we directly test using the `podman-tests`
package. This means we'll have to disable podman e2e tests and only do
system tests but that shouldn't be too big a problem.

A better way to run podman e2e tests would be by creating a `podman-src`
rpm subpackage that'll install all the rpm sources.

The same set of tests can be run across all environments so there's no
need to maintain separate plans for upstream and downstream.

Signed-off-by: Lokesh Mandvekar <lsm5@fedoraproject.org>